+++ src/view/CadastroSessao.fxml
@@ -8,21 +8,21 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.CadastroSessaoController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.MenuController">
    <children>
       <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="410.0" layoutY="319.0" pickOnBounds="true" preserveRatio="true">
          <image>
             <Image url="@../imagens/footer-pipoca.png" />
          </image>
       </ImageView>
-      <Label layoutX="66.0" layoutY="38.0" text="Formulario cadastro de Sessao">
+      <Label layoutX="66.0" layoutY="38.0" text="Formulario cadastro de Sala">
          <font>
             <Font name="Arial Bold" size="19.0" />
          </font>
       </Label>
-      <TextField fx:id="inputSessaoNome" layoutX="178.0" layoutY="90.0" />
+      <TextField fx:id="inputSalaNome" layoutX="178.0" layoutY="90.0" />
       <TextArea fx:id="textareaSessaoNome" layoutX="177.0" layoutY="135.0" prefHeight="171.0" prefWidth="399.0" />
-      <Button fx:id="btn_sessao_salvar" onAction="#cadastrar" layoutX="179.0" layoutY="334.0" mnemonicParsing="false" text="Salvar" />
+      <Button fx:id="btn_sala_salvar" layoutX="179.0" layoutY="334.0" mnemonicParsing="false" onAction="#cadastrarSessao" text="Salvar" />
       <Label layoutX="109.0" layoutY="95.0" text="Nome:" />
       <Label layoutX="85.0" layoutY="142.0" text="Descriçao:" />
    </children>
