+++ src/view/Menu.fxml
@@ -14,8 +14,6 @@
               <menus>
                   <Menu mnemonicParsing="false" text="Aplicaçao">
                      <items>
-                        <MenuItem mnemonicParsing="false" onAction="#inicio" text="Inicio" />
-                        <MenuItem mnemonicParsing="false" text="Gerenciar FIlmes" />
                         <MenuItem fx:id="btn_filmesEmCartaz" mnemonicParsing="false" onAction="#filmesEmCartaz" text="Filmes Em Cartaz" />
                         <MenuItem mnemonicParsing="false" onAction="#sair" text="Sair" />
                      </items>
@@ -26,15 +24,14 @@
                         <MenuItem mnemonicParsing="false" onAction="#cadastrarSala" text="Cadastrar Sala" />
                   </items>
                 </Menu>
-                  <Menu mnemonicParsing="false" text="Filme">
-                    <items>
-                      <MenuItem mnemonicParsing="false" onAction="#cadastroFilme" text="Cadastrar Filme" />
-                    </items>
+                <Menu mnemonicParsing="false" text="Filme">
+                  <items>
+                    <MenuItem mnemonicParsing="false" onAction="#cadastroFilme" text="Cadastrar Filme" />
+                  </items>
                 </Menu>
                 <Menu mnemonicParsing="false" text="Sessao">
                   <items>
-                    <MenuItem mnemonicParsing="false" text="Gerenciar" />
-                        <MenuItem mnemonicParsing="false" onAction="#cadastroSessao" text="Cadastrar Sessao" />
+                    <MenuItem mnemonicParsing="false" onAction="#cadastroSessao" text="Cadastrar Sessao" />
                   </items>
                 </Menu>
                 <Menu mnemonicParsing="false" text="Configuraçoes">
